This PR enables running in unpriviliged mode when run as the splunk user account. The splunk user is now part of the ansible group and able to run the ansible playbooks itself at startup, but unlike ansible, it has no sudo capabilities at startup time.
Note that you currently must explicitly set SPLUNK_HOME_OWNERSHIP_ENFORCEMENT
to false when running as the splunk user, otherwise it will fail due to
lack of permissions.

Note that there are limitations, beyond just being unable to "correct"
permission for volume mounts. Any features requiring elevated permissions,
such as install JDK or other packages, will not work when running as the
splunk user.

The defaults are left unchanged. By default, it will still run as the
ansible user and the behavior should be the same as before.