add btool check for HTTPS on health check (#488)
Not all methods of disabling HTTPS on the API will set the SPLUNKD_SSL_ENABLE flag, which would then cause this health check to fail, and the container to restart.  This change checks the effective enableSplunkdSSL value in server.conf, which is set to "false" if defaults.yml has ssl.enable set to 0.